fix(ui): wire series adapter end-to-end and disable caching; normalize DTP3 to {year,value%} for chart rendering; add server log proof
