+++ package.json
@@ -20,6 +20,7 @@
     "repair:dtp3": "tsx scripts/repair_dtp3_schema.ts",
     "verify:dtp3": "tsx scripts/verify_dtp3_schema.ts",
     "debug:dtp3": "tsx scripts/debug_print_first_item.ts",
+    "peek:dtp3": "tsx scripts/peek_dtp3_ui.ts",
     "pipelines:firearm_stock_per_100": "node scripts/pipelines/firearm_stock_per_100.js",
     "validate:sources": "node scripts/validate-sources.cjs"
   },
